Patch by pieterg:
Some providers manage to send epg which matches on the bouquets, in the
bouquet list. (probably sid 0 or something)
As a result, nonsense event info is shown in the bouquet list.
And because the bouquet names have a nonzero offset (because of the 'folder'
icon), the bouquet name overlaps this event info.

The simplest workaround is to only show event info for 'playable' items,
which might even slightly speed up showing the bouquet list btw.