Patch by Moritz Venn:
The current implementation of ConfigNumber.isChanged (the one inherited from ConfigElement) does not work properly as it will - at least for unsaved values - always returns True.
This is obvious if you just take a look at the datatypes since value is an int and default is a string (because ConfigNumber is a modified ConfigText).

To resolve this issue one can either compare self.tostring(self.value) or self.text to self.default but the former seems to be the more logical approach.
The attached patch does override the method in ConfigNumber with the proposed fix.