By Anders Holst:
At regular intervalls there are questions in the forum (in at least
three or four different threads on only this subject) on how to
configure the recording paths so you don't have to set them to the
same thing over and over again in e.g EPG timers. People apparently
forget to set it, and miss their recordings because the harddisk is
full or not built in. They want an easy way to change the default from
/hdd/movies/ to a NAS or a USB device.

I have rebased the patch in this thread from March 23, added two minor
bugfixes, and hopefully made it less controversial by making sure that
it only affects the expert setup level. I don't think there should be
anything controversial about this patch now, and no known bugs. It
only provides functionality that several people are screaming to get.

Synopsis: The patch changes the "Timeshift path..." setup dialogue
into a "Recording paths..." dialogue where you can easily configure
both the timeshift and the other recording paths.