Assume cable and terrestrial ONIDs/TSIDs are always valid
We were checking them against the (satellite-specific) blacklist.
As a result, the namespace field for cable and terrestrial services
got unpredictable values, depending on their ONID/TSID; sometimes
the frequency was included in the namespace, and sometimes it was not.

This patch fixes the problem by assuming ONID/TSID is always valid
(and unique) for cable and terrestrial services.