Patch by Moritz Venn:
"I've been annoyed by the incomplete "sort by date" for a long time - it only works for recorded movies.
Not much to say, I added a simple stat call in service.cpp and return the mtime for sTimeCreate so the list is as properly sorted as the dreambox can ;-)"

Minor changes so that it is implemented in servicemp3.cpp, since a service reference "path" is specific to the actual service.