fastscan: don't obey dxHoldName flags when useFixedServiceInfo is set
We used to obey dxHoldName, leading to the situation where we would set dxHoldName
ourselves, and then in the next run refuse to modify the service names because
we found the dxHoldName flag set.
To avoid this unrecoverable situation, we choose to ignore the dxHoldName flag,
whenever the fastscan is run with the 'useFixedServiceInfo' option.
In fact, the user has explicitly granted us permission to modify service names
in that case.
The default is still to run fastscan without 'useFixedServiceInfo', so fastscan
will never modify any service names, unless you tell it to.