ChannelSelection: merge the new 'Activate Picture in Picture' option with the existing one
We already had 'play as picture in picture' and 'play in mainwindow' options in the
channel context menu.
And those are properly obeying the parental control, the new on was not.
Added blue key shortcut to the 'play as picture in picture' option, and use the new
'showServiceInPiP' instead of the old 'playPiP'