keymapparser: handle generic key entries before device specific ones
This ensures we'll show the best match in the help screen.
Generic keys are usually keys with only a single function,
e.g. KEY_PAUSE for pause. If a remotecontrol has this key,
we prefer to show that one in the help screen, rather than
a yellow key for instance.