teletext subtitling fixes by littlesat
Positioning and coloring are kept as close to the teletext
information as possible.
However, all individual lines are centered.
One issue which still needs to be fixed: when multiple colors
are used on a single line, in order to let the painter do the
centering, we have to resort to using the same color for the
entire line.