VirtualKeyboard: fix backspace on non-ascii characters
Backspace in utf-8 data would break utf-8 sequences, because it did not
clear the whole multibyte sequence at once.
So keep the text in unicode internally, do not encode to utf-8
untill we have to display or return the result.
And decode the initial utf-8 text we receive in the constructor back
to unicode.