VirtualKeyboard: fixed querty layout, and use line lengths of 12 keys
The cursor navigation routine is currently assuming the first three
keyboard lines to have 12 keys, where the querty layout had a few
lines of 11 keys.
This could be fixed by using the actual lengths instead of hardcoded values,
however adding a few keys to get 12 per line is a lot quicker, and will do
for now.