eTextPara::renderString: avoid running out of string bounds
When the last character in the rendered string was a '\', the
escape code check caused us to run past the end of the string buffer.
Usually, this would result in an endless loop.
Added a sanity check to avoid this.
Note that with the current escape code implementation we still
risk getting unexpected results, when strings contain '\'. But
at least it should not crash anymore.