font rendering: added support for borders
When we want to render borders, we cannot use the simple bitmap cache
anymore, so instead we will use the (less efficient) image bitmap.
To compensate for it being less efficient (and the fact that
we have to convert the image to a bitmap, after retrieving it from
the cache), we store the bitmap in the glyph vector untill we have
rendered it.