bsod: do a clean shutdown after a python expection
Don't let the C++ stuff suffer for python bugs ;)
When a python exception occurs, we can assume that the
c++ objects are still fine, so we prefer to do a clean
shutdown.
As a result, for instance the channel database and epgcache
will be properly saved when we hit a python exception.
(python exceptions are the most common cause of a bsod)