eTextPara: fix rendering with equal foreground and background
We were setting the current foreground color to the background,
to force an initial color change.
However, when foreground and background were the same (don't ask)
we failed to initialize the color lookup tables, and we would
segfault.
So we have to use a different hack to ensure we see an initial
colorchange.