eServiceMP3: fix the race on the GstBus message handling
We simply don't feed the messages into the bus anymore,
instead we put them into our own messagepump, and tell
gstreamer to drop them.
On the other end of our messagepump, we handle and unref
them.

This works if we assume that we are the only one interested
in those messages, and no internal gstreamer messages are
posted in 'our' bus.
So let's try this for a while, and find out whether there are
any sideeffects.