Clean Trashcan while preparing recording.
There's no event to hook "prepare" to, so the Trashcan
is hardwired into RecordTimer.
The prepare happens a while before the recording actually
starts, which is a perfect moment to clean out the trash
and make room for a new recording if diskspace is running
low.

Also a few refactorings, such as removing 'static' methods
from the Trashcan class.