Experimental posix_fadvise for recording
Instead of sending something totally random to posix_fadvise (because
some header is broken), tell the kernel what we intend to do with
the recording file.

Recording one or two HD channels to NFS appear to run "smoother" with this
setting, which may be coincidence but at least it's better than randomly
dumping some stack contents to a kernel function.