dm8000 input device: fix play/pause key event
The dreambox advanced remote control has a combined play/pause key.
However, it emits a KEY_PLAY event, instead of the expected KEY_PLAYPAUSE event.
Working around that in the keymaps causes problems for input devices with
real PLAY keys, so instead we translate the key event to KEY_PLAYPAUSE, before
reporting it.