Another seek fix by Aholst from the enigma2-devel ml ------------------------------------------------------------------ And one more fix on this theme: rewind and higher than 8 times forward doesn't work for some movies.
It turns out that when I and Felix communicated about the "accurate
speed winding" that he eventually implemented, introducing the .sc
files, we both made a false assumption (sorry if I lead you into it),
that it was enough to jump to an I-frame picture start. However, it
appears that (for an SD-movie that is) one should jump to the sequence
start instead. (This is actually what the c++-code already does when
dealing with the .ap-file, but not for the .sc) Since most movies have
the sequence and picture start in the same ts-frame it usually doesn't
matter. But there are movies out there with the sequence start in a
ts-frame before the picture start, and for them rewind and very fast
forward currently just freezes playback instead.

The below patch fixes that. It just backs up through .sc to the previous
sequence start once the picture start is found. Also, alignment has to
be done here again and not in dvb.cpp, to get the right winding
speed. Please ask me for the full technical details if something is
unclear.
------------------------------------------------------------------
refs bug #570