Cache data in MovieList for faster display
Instead of opening files on the harddisk for every bit of info, just
remember what was returned last time and store that into the list,
like was being done for (only) the length of the movie.
This should speed up the movie list display considerably.

There's still some debugging code in there in case this causes trouble
somewhere.