fsck runs in background, fix it on 'disc' mounts
Developed and tested on 7025 built-in harddisk (1 partition) and a
CF card (no partition). This now works correctly (this was
broken in E2).

Tested on VU+ with a USB stick.

But I still may have broken things in some unexpected way.

TODO: Add notifications and display errors on the UI. The
feedback is really bad currently (worse than before).