MovieSelection: Stop recording from within movie list
If recording, recording will be stopped, repeating timers will be cancelled just once.
If the file appears to be in the progress of writing, you get a warning (file busy),
for example when another box is recording to a NAS.

Todo: update the red button text and offer choices like "stop and delete", "stop only".