pmt: work around refcount issue with backgroundscan disabled
We used to not create an eDVBScan object, when background scanning
was disabled.
However, this somehow caused the channels on other transponders
to be greyed out in the servicelist, indicating some kind of
channel or demux refcount bug, triggered by not starting the dvbscan.
So as a workaround, we now always start a dvbscan, but choose to ignore
the result when background scanning is disabled.

We should actually find out what causes the problem, and fix it.
But untill then, this will have to do.