Integrate DVD playback into Enigma2
Experimental - should allow DVD playback without plugin in time.

For now, you DO need the plugin because it needs to do slightly
more than just start the service. This may change in future.

The immediate gain is saving ~500k flash because the DVD plugin
also installed .a and .debug files.