Use colors instead if italic and bold for pango subtitles
As this is the quickest work-a-round as italic and bold are not
used much in pango subtitles. For correct italic and bold
subtitles we need a huge patch instead. Italic will be displayed
as cyan and bold as yellow. Later the subtitle fonts with faces
could be stipped as this code is not required anymore.