db: add support to retrieve a list of servicerefs
The database could contain duplicates with different servicetypes.
This happens when a service changes its servicetype for instance.
In some cases, we do not want to find just the first matching
serviceref, but all duplicates as well.
Because the first one might not be the most uptodate.
Note however that searchAllReferences will take twice as long as
searchReference, on average.