skin: Fix LCD/OLED disappearing without Magic
Was caused by a missing windowstyle in skin_display.xml
Added it to default, because other stuff may expect it to be present
even if no such display exists. The one in skin_display is the one
to rule them all and in darkness bind them.

Also removed unused skin_display_dream.xml