Add option to swap SNR in dB and percentage on OSD
As some users want to see a percentage on the usual OSD
and this avoid that users need to change a skin.xml.
I choose for an swap option as on some screens (on some
skins) both a value in percent and dB is displayed and
a swap option makes it a simple yes/no option.