Fix for DVB grafical subtitles during recordings when PTS lakes
This will fix issues while playing recordings or timeshifts from Travel Channel HD.
Please note that in fact the PTS on this channel is corrupt. A PTS lakes when
there is a diffence of more then 20 seconds (which was always the case in enigma2
even without playing recordings). The same config is used as when the teletekst
subtitles do lake during playing of recordings and/or timeshifts.