Hotplug: also close thread on language change
Previously after a language thread the bdpoll variable was overwritten
leaving the old thread run forever and only closing the new one on
shutdown (a language changes does NOT call the plugin close handler but
the start handler is called again). so before creating a new bdpoll
instance check if an old one exists and unset the running flag.
Signed-off-by: Moritz Venn <moritz.venn@freaque.net>
Signed-off-by: ghost <andreas.monzner@multimedia-labs.de>