picload: Load 8-bit images into 8-bit surface
The GDI can display 8-bit palette images really fast, so converting
them to 24-bit and then to 32-bit is a terrible waste of memory and
CPU.

The background sucks for now (just uses index 0 of the palette), but
rendering and loading is much faster.

TODO: Treat the background setting nicely (find matching color, or
insert one into the palette if there's room for it).