tstools getPTS: Fix and speedup "resync"
Created a TS recording file, prepended 20 zero bytes to it to misalign it. Much
trouble playing and listing it. Created a .TS file with just zeroes, which
causes short "hangs" to occur when displaying a recording list.

Apparently the offset was calculated wrong, it was looking for the 0x47
sync marker, but if found at position X, it would set the offset to
188-X instead, and then repeating the process.

So look for the 0x47 byte (using memchr is way more efficient than a loop)
and adjust the offset accordingly. The 20-byte misaligned file now plays
smoothly and doesn't output "resync" a hundred times.

Another optimization is to just give up when resyncing over 64 times. Just
assume the file is corrupt or so. That fixed the "hangs" while attempting
to parse the all-zero file.