Remove only show second infobar when OK was pressed twice within 750 msec
As the within 750 msec condition is tricky as my interfere with other functions
like long press key and double click keys what is "arranged" deeper into E2.
I think this is also not needed as the second infobar without is working as
before and for those who do not like the second infobar can disable is which
is even the default setting now.