pvrparse: Separate structure and accesspoint availability (experimental)
When playing a recording in progress, the "sc" file is up to date but
the "ap" is missing. Just use the sc file, it's there, you only miss
the last 4k (which is usually ~4MB of recording data).

The effect is that you can now smoothly (re)wind through recordings
while they're still running.