Change default for mask for auto language subtitle equals audio language
As this did confuse users. Usually you expect that a prefered subtitle will
always apear, but as it was and the audio language was the same we see no
subs at all and if you wish this you need also change the mask. By changing
the mask to all then always the subtitle will appear by default. When you
do not want this you still have the possiblity to change this mask as
before.