pvrparse: Use os filedescriptor to write sc file
Prevents too much caching in FILE class. By writing out in PAGESIZE chunks
the efficiency should be at least as good as FILE does, but it prevents
updates to the sc file taking way too long.

This may fix reported issues playing back recordings while they are running
on remote filesystems. At least I've noticed that when recording to an NFS
share, nothing was written to the sc file even minutes into the recording,
whereas when recording to a local disk it flushed every 4k.

(TODO: Get PAGESIZE from kernel stuff. It's another step on the road to
mmap for this stuff)