pvrparse: use mmap to read and write sc data
The writer mmaps a shared page block, and updates it as the data comes in.

The reader mmaps portions of the sc file, and can thus see the live data
coming in too, because this will point to the page that the writer is
accessing.

The result is that the parser can now see every frame as soon as it's been
written to disk. In the UI you can see that the total time of a recording
no longer jumps in ~4 second increments, but keeps a steady state. The
lenght is also recalculated much more often, because it is now cheap to
to do so, even without .ap file.

(With lots of logging, to be removed later)