MovieSelection: Remove seek setting, start at movie, letters first
When the movielist is opened, go to the first file.
Remove the setting for behavior of number keys. There are so many settings already, so I'd rather keep it simple. Who'd want the old behavior anyway?
When pressing the "play next" key, start a preview. This solves the problem that not all remotes have a play button for preview.
Handle the letters first, numbers later.