servicemp3: Add option to stream to enable/disable prebuffering
Because for most streams this is not needed as simply the default
buffering is working. The prebuffer will no only be used when buffer=1
is added to the URL. There is still some investigation needed regards
this buffering. This avoids the shoutcast and waiting issues for the
common streams. Probably if we want to go this way we have to find
a better solution for the buffer=1 trick as it is not nice especially
when per accident the same sting exists in an URL.