eDVBServicePlay: fix isCurrentlySeekable logic
When we did not yet have a valid decoder pointer, isCurrentlySeekable
would return 0.
For TS playback without a meta file, we do not fill the decoder pointer
till we found and parsed PMT from the stream.
Usually, isCurrentlySeekable is called (by the infobar) before we
find the PMT. As a result, seeking was not allowed.

Fix the logic, which was merely intended as sanitycheck for
m_decoder->getVideoProgressive(), allowing seek for TS playback
when the meta file is missing.