Non-recursive make part 3: Allow parallel build of SWIG and the C++ libs
The SWIG compiler and the files it generates are a major time-consuming part
of the build. When in a "subdir", the whole build will wait until a single
thread has finished building this. After this modification, the swig part will
run in parallel with the other C++ files, which is "noticably faster" on
multicore systems.