demux: Use userspace buffering in favor of demux size
The IO buffers "push" the IO on the system much harder, the demux
buffer is not under "our" control. The IO buffers are now chosen
based on installed memory size instead of the demux size.

Formatted a 160GB disk in the et9x00 as ext3, and mounted with the
ext4 driver (data=ordered). Started 8 simultaneous recordings on the
box (6xHD 2xSD). While these were running, use dd to write 2GB files
on the same disk, erase, and again. This resulted in no overruns.

Behaviour on a slow flash stick also improved, less overruns, and
dd doesn't cause overruns either.

My conclusion: Pushing more data to the OS (instead of holding it in
the demux buffer) causes more "pressure" and thus helps the OS push
the data out at higher rates.