decoder: detect whether or not we can rely on VIDEO_EVENTs
Some hardware does not invalidate the video attributes,
when we open the video device.
If that is the case, we cannot rely on receiving VIDEO_EVENTs
when the new video attributes are available, because they might
be equal to the old attributes.
Instead, we should just query the old attributes, and assume
them to be correct untill we receive VIDEO_EVENTs.

Though this is merely a cosmetic issue, we do try to detect
whether attributes are invalidated or not.
So we can avoid polling for valid attributes, when we know
we can rely on VIDEO_EVENTs.