Language: set LC_TIME locale when langua is set/changed
When the language is set/changed in the system menu the system date
functions can also work in the new language by setting the LC_TIME
locale. For this to work the specific language needs the LC_TIME
file in the system locale directory.
This will be done in a seperate update.