RecordTimer: Prevent recording to rootfs, create missing dirs
When preparing, attempt to create any missing directory. This automatically
creates /media/hdd/movie for example if it was missing.
Check the mountpoint, and dont start recording if that is the root. This
prevents recording into internal flash when the harddisk isn't mounted.

The combination of the two is that you can format the harddisk while a timer
is attempting to start.

Also sanitized the Directories module a tiny bit.