dvb: avoid a race counting frontend device nodes
When we link a vtuner device, a new frontend device is added.
The devicenode however will not exist right away, as it will
have to be created by udev/mdev.
So when counting frontends to detect which will be the next number,
it is better to look for sys entries.

Theoretically, we have the same race in the eDVBFrontend
constructor, opening a (virtual) frontend device.
However, it looks like mdev has created the devicenode before
we get there eventually.
If it turns out we are hitting the same devnode race there,
we will have to add a retry loop in openFrontend.