frontend.cpp: Add NovaT 500 stick and change max snr for dvb-T to 29 dB
Principally we should use for another solution for the snr to db calculation
e.g. the drivers should forward always the same standard values as now for
for each type of USB tuner we must add something in frontend.cpp. But for
now at least I added the reported NovaT 500 stick. For other USB sticks
we also did not receive any reports yet.

Also changed the max snr in db level to 29 dB as this is the real max
that could occur so we see no percentages above 100% anymore