SecondInfoBar: A diffirent way to save resources.
When the second infobar was hidden the PliExtraInfo was not polled anymore
except when enigma2 was started and the secondInfoBar was never showed
before. By using onLayoutFinish.append we can toggle the secondInfobar
being showed and hidden at initialization. This will save receive resources
also when the SecondInfoBar was never shown before.

This will also allow to use PliExtraInfo outside a SecondInfoBar

There can still be more resources be saved later by not "polling" frontend
data and service information when it is not requred in PliExtraInfo