pliExpertInfo: make it obsolete as PliExtraInfo is the successor of it
pliExpertInfo. The successor is  more efficiant regards functionality,
structure and use of resources.

Please note that It could happen that some skins can get blank screens when
they use pliExpertInfo - so they need to addapt to PliExtraInfo what is
recommended -or- they need to include pliExpertInfo in the ipk package