tstools: Fix fast-forward of MPEG2 streams at >16x
The "backsearch" sometimes would cause the stream to go back in time.
This is now prevented by limiting the search for a start frame to the
number of frames skipped. Also reports the correct number of frames
back to the caller, and a tiny perfomance optimization by moving the
"if" outside the loop.
The limited search also applies to rewinding, which isn't entirely
correct.

Problem was introduced in 4301e468e6c4f530e85e3aaa1b2a89b3235a4487