Add parentel rating and content description to now/next events
event.cpp only seems to deal with no/next events.
The parental en content descriptors allow for multiple
definitions to exist at the same time, so lists of information
is maintained at low level. In practice it seems only one
parental and one content definition is entered by the
broadcaster.
The interface to python will only pass the first level.
On python level the tranlsation to a readalbe format is done.
A simple implementation is added and is subject to change...

Updates for epgcache will follow.