picload: allow to resize without keeping aspect ratio
Resizing always tried to keep aspect ratios.

With this change when an aspect ratio is given as zero, e.g in:

setPara(width, height, aspect_x, aspect_y, cache, resizetype, bgcolor)

where aspect_x and aspect_y are zero the resizing will just fill the
widthXheight area with a scaled version of the pixmap file.
This can for example be used to stretch a smal image to create a full
length bar or column of any size.

NOTE: in future when aspect_x is zero the meaning of
      aspect_y may become different. The idee is to allow for
      different types of filling areas...

Also avoid a division by 0 possibility.