GraphMultiEPG: Change font for service name when only picons
When only picons are displayed the channel name is displayed in case
for a service there is no picon. Now choose font = 1 which us the same
as the event font but usually smaller then the standard service font.

In addition the service name is rewrapped to allow more lines and
it is centered for better looking