Navigation: Fix zapping along services with alternatives
When zapping along a service that had an alternative we were zapping
incorrectly and even the channel number was incorrect. This was intro-
duced since I added the setcurrent call to the service list. This
was caused due to the fact some alternative ref could be stored there
instead of the orriginal reference.