ServiceScan: Derive infobar from session and fix a possible GSOD
Infobar should already be available via session so there is no need
to import it to get access to the servicelists
In addition self.session.postservices was declared in ServiceScan and
when we were in ScanSetup without doing a scan session.postScanService
does not exist. Now session.postScanService is declared in ScanSetup.