Completely revise the channel numbering
Based on the suggestions made from Awk in this thread:
http://openpli.org/forums/topic/21514-service-number/

Currently it is doing exaclty then same as we were used to.
So (hopefully) nobody experience changes.
But in the code we do not have to count back for markers and
we do not have to take track (determain) the first service number
in a bouquet. So this might to save resources.
within the service list there exist a number field that
will be filled during reading the bouquets and after a change
it can be renumbered.
Possibly later the bouquet number cache could be removed
or stuff like hiding number markers could be added aswell.