GraphMultiEPG: Allow to exit after zap
For close False en True was used. Previously True was a recursive close
but I change that to restart the GME to switch modus (list/pig mode). Per
accident I missed that True was already used for recursive close. This
commit corrects that by using None instead of True when restarting the
GME is required for e.g. a different mode. And also the recursive is also
correctly "forwarded". I also removed the variable self.closeRecursive as
due to the new Zap/Preview functionality it is in fact not required
anymore