InfoBar: Add confirmation before exit Movieplayer
As there were conserns that when press exit in the movieplayer
that you suddenly stop the movie player I added a confirmation
simple message box. With selecting No or simply press exit again
you get do not stop the movie. When selecting yes or simply press
OK you will still exit the movie (as yes is default).

This avoid a new configuration setting and adds functionality so
that you also can stop a movie with the exit button which did
orriginally nothing there.