eDVBTSRecorder: increase demux buffer size to 1MB
The (in-kernel) demux ringbuffer does not need to be
packet aligned, we can pick any size.

And even though our userspace buffers are quite large,
we need a reasonable kernel buffer to cope with fluctuations
in the rate at which userspace is able to read (and handle)
the demux data.