Add option to display tuner type in case of no orbital position (DVB-S)
When it is DVB-S then the orbital position is given on the OSD and the
channel list. When it is DVB-T or DVB-C nothing was given. Now we have an
option to give DVB-T or DVB-C instead when there is no orbital position
For the PliExtraInfo converter and also for the ServiceOrbitalPosition
converter we have the extra option OrbitalPositionOrTunerSystem that can
be used in skins for this (see examples in standard PLi-HD skin)