MovieSelection: Simplify playlist handling
It looks like a list, so use a list instead of a new class.

Make MovieList behave like a list, so you can iterate through it instead
of having to use a loop counter and getItem.

A few minor cleanups too.

I don't like the global "playlist", but didn't quickly see a way around
it.