servicedvb: fix audio pid cache entry
When we have only one audio stream, we can safely overwrite
the audio pid cache, without risking to remove any
user preference.
Often, the audio stream type might be unknown, or stored
into the cache incorrectly (while scanning).

This modification gives us the chance to fix the audio
stream cache information when there is a single stream.
(if there are multiple audio streams, the user will
cause the cache to be fixed when selecting an audio
stream from the audio selection list)