Language: add ".UTF-8" where os.environ['LC_TIME'] is set
In the previous related commit we could not force UTF-8. But with
adding a dot and UTF-8 after the language it fallsback to UTF-8.
I verified it with different languages as fallback to check if the
fallback is indeed working as when trying it with ".ISO8859-1" it did
not fall back to ".ISO8859-1". In addition I discovered that when
invalid data is entered into os.environ['LC_TIME'] python does not
use this fallback en then you also do not have a language change.

As I tested that the fallback with .UTF-8 works for different languages
in os.environ['LC_TIME'] I suggest this is a good working solution.