Movielist sorts: add flat sort mode
The flat sort option allows to sort alphabetically without
differentiating between files and directories. Thsi can be very
useful if you have a directory with simple filebased movies
intermixed with subdirectories that contain for example episodes
of a paricular television series.

The only thing is that the default sort order is based on
last modification date (e.g. recording date). This may only
be useful in recording locations but not for your collections
of movies or music. The sort can be defined per directory but
it is quite a hassel to do that for all directories while there
are only a few recording locations.
So the choice is to just change the default sort order and explicitely
set it for recording locations if you want sorting on recording date;
or add a user interface option to specify the default sort option....