Satconfig / AutoDiseqc: remove 'auto' mode
Work with the existing 'simple' mode instead;
autodiseqc is activated when at least one
of the ports is configured to be 'automatic'.
Allow auto detection in a/b/c/d, a/b as well as single mode.
(resp. 4, 2 or 1 port detection)

Always return to the configuration screen after autodiseqc,
regardless of the result.
This allows the user to check (and possibly correct) the
detected configuration, or restart the auto detection.

All ports are set to 'automatic' by default, instead of
the (randomly chosen) 19.2 / 13.0 defaults.