timer: keep short timer activation poll interval, when system time has not yet been set
When the system time is not yet valid, the next activation interval
cannot be calculated correctly (it will be several years in the future).
The interval is limited to 100s (MaxWaitTime), but this is still a rather long
time when recordings have to be started right away.
So instead, keep a short (5s) poll interval, while the system time
is before 01.01.2004.