filepush: Reduce overly complex filterRecordData and a few others
Since filterRecordData is defined in only one spot, and just to peek
at the data without ever messing with the buffers, simplify the
function's signature, and reduce the handling code to match.

Move the buffer start/end variables to local variables, which removes
the need for a flush() method as well.

Remove large blocks of dead code from dvb.cpp's filterRecordData.