filepush: Faster "stop" and remove unused start() methods
Only one of the start() variants was ever used, removed the other two.

Made the filepush loop more logical: Read first, then write data.
The stop() method sends a signal, which interrupts the read/write in
progress. So after a failed call to write() or read(), check the m_stop
flag first and abort immediately. This should make skipping and so more
responsive.