rawfile: Set POSIX_FADV_SEQUENTIAL
The only real effect in the kernel is that the read-ahead page count
is doubled. The readhead max is 128kB (VM_MAX_READAHEAD) so doubling
that to 256k should be benificial.
The advise is not used when opened in "cached" mode, as this suggests
that random access is expected. Only tstools.cpp does this, and only
when looking for PTS data in the raw stream.