Add doxygen documentation system
Just a basic placeholder directory with a README file
to briefly explain how to use this.

The doxy.gen file is a doxygen configuration file with
a basic setup for getting class relation graphs for enigma2
c++ and python files. This probably is far from perfect so
suggestions to updates this config file are welcome!

Maybe this can lead to a better documented enigma2. Don't hesitate
to post enigma2 documentation patches using doxygen format...