Navigation.py: Introduce getCurrentlyPlayingServiceOrGroup
Instead of introducing a 'selected' argument keep the original
'getCurrentlyPlayingServiceReference', which always returns just
what is says, the currently playing service reference. This also
allows backwards compatability with possibly some plugins.

Introduce a new function, 'getCurrentlyPlayingServiceOrGroup', which
returns the current alternatives group (alternatives) in case the
current service belongs to a group, or the service itself in case the
service is not a member of such a group (alternatives)