tstools: Correctly and efficiently calculate recording length
A recording with dicontinuities, such as when you've cut out commercials
or when there were big dropouts during recording, did not account for those
when the length was requested.

To fix this, take the discontinuities into account, which are calculated from
the .ap file. Use the "fixupPTS" function to correctly estimate the length.

Improve efficiency: Keep the calculated length as a member, because it is being
requested many times per second while playing.