InfoBarGenerics: Some tiny timeshift related fixes
Fixed when going to MovieSelection and directly exit that timeshift
could be started, but not played anymore.
Remove PVRState when zapping/selecting different service that also
stops the timeshift.

I still discover some strange things in case the timeshift buffer
does not have much time yet (< 1 min) and we step skip back. This
is not fixed (yet)