[skin] add some code so that skin.xml files can detect there folder, so that we can now use relative paths, this makes it possible to allow includes to be truly useful. see below for an example.
if we use pixmap="screens/fs-bg-top.png" it first looks in default location (/usr/share/enigma2/) for "screens/fs-bg-top.png", if it does find the file it then looks in current active skin folder i.e. (/usr/share/enigma2/ViX-Night-HD/) for the same file. (/usr/share/enigma2/ViX-Night-HD/screens/fs-bg-top.png).

but where this really useful is in includes, because as the skin.xml is the ACTIVE skin it tries to look for the file in there thus allowing you to the same skin to load different images locations.

See ViX-Night-HD & ViX-Day-HD for true examples. these skin are identical apart from skin colours and image files. so we do not need to maintain 2 full skins only one include file.