Protect timeshift at least during normal zapping
But still the movieplayer, scanning for (new) channels and a lot of
plugins might abort a timeshift in progress. Usually any time when
stopService is called the timeshift will be stopped. I could not find
a general way to protect a timeshift in progress. At least I hope
that the function I did create in InfoBarGenerics could be used in
some kind of universal way in case the timeshift protection will be
extended to other functionality.