vtuner: disable dynamic PES filter start on stb's with older DVB API
following cd64e9661ab364804b4344051bcf0716cc8b1992, it turns out
that on several stb's the stability of the vtuner interface
was dramatically reduced, now that we are dynamically starting
and stopping PES filters, as they are needed.

So for older DVB API versions, in order to work around buggy
vtuner implementations, we pre-start a PES filter, making sure
'pidcount' never drops to zero, so the filter is never stopped.