PicturePlayer: add option to configure full view resolution
Usually pictures were displayed on the same resolution as the current
skin. Now the resolution of a picture been shown can be configured. During
fullview the desktop overlay will be temporarily be switch to the
configured resoltion during the picture been shown.

This makes it possible to view the picture in full-HD even when your skin
is not in full-HD.