Revert "PicturePlayer: add option to configure full view resolution"
This reverts commit 0a0d9fc19398804e96a35086974e25fb95c6081b.

Changing the resolution leads to a re-assignment of the graphical
acceleration buffer making all accelerated pixmaps invalid when they are
removed or replaced (picons).

After a resolution change you are not able to zap anymore because you get
a GSOD at the moment memory is made free for the Picon.