Change blackfill color to transparant black for scalable picons
When picons do have a (8 bit) color map without any transparant colors we
had a white blackfill when the picons are scaled and centered into the
destination. Hopefully backfilling in transparant black will be a better
work-a-round. However for correct displaying it still should be considered
to use non (8-bit) color mapped picons.