gdi/gpixmap: Don't use broken hardware alphablending
Alphablending on (at least) the ET boxes is broken. When one of the
pixels is fully transparent, the resulting pixel is always fully
transparent instead of simply the other pixel's argb values.

This resulted in objects not being rendered at all, when the background
is transparent. It is also the reason why a transparent picon will occur
as a "hole in the UI" instead of just being invisible.

As a workaround, don't use hardware alphablending (even if the hardware
reports it as a capability) except when scaling. Historically, this has
been the only use for hardware accellerated blending anyway, so there's
no loss in performance.