More realistic hardware blit usage policy
Some quick measurements indicated the price of setting up a hardware
blit was quite high. Rendering large pixmaps is considerably faster
though (70ms vs. 6ms on 1920x1080 surfaces) so limit its use to only
larger surfaces, and avoid it on smaller ones.

On 8-bit images, alphaTest blitting works okay. On 32-bit it appears
to be broken, so don't use it.