listboxservice: do not keep aspect ratio anymore
As the target has already the expected aspect ratio keeping the
aspect ratio is not required anymore. So currently it can always
be scaled without having blackfilled borders.
It might have  a disadvantage with picons that big differ from the
usual picon relations.
In addition add a margin of 8 to the service name so the service
name is not "glued" to the picon.