Simplify and speedup the LanguageSelection Screen
As discussed internally we do not upgrade the title to the current
selected language anymore and the same for the instruction on the Wizard.
We just show it in the language the Screen was initiated.
For the user interface upgrading titles and instructions  just looks nice
but principally it has no real added value.
This will increase the response this screen because with each time a new
language was selected we do not need to read a complete language file
anymore. And there are also no hacks required that might cache and/or store
in a "work-a-round" way.