Visually change behaviour of FIFO list to as it was before
Visually it behaves as it was before and will show only the latest len
items that were added to the list. When listAll is called the complete
list will be shown. So during a scan everything was like it was before,
but after the scan you're able to select a specific service.