ServiceScan: Fix for restarting startup Wizard
When a service scan was performed during the startup wizard after a
service scan the wizard was restarted. This was caused by the recursivaly
closed functionality added. Now when there is no basic Infobar the
servicescan screen will be closed without any parameter. This allows the
setup wizard to behave normally as before.