Add OpenMP support (picload only for now)
Getting OpenMP in OpenEmbedded was step one. Step two is to actually
use it in Enigma2 code to conveniently run things in parallel. As an
introduction, apply it to the image resizing functions in picload.

Other areas of interest would be initialization code, the EPG search
methods, and the GDI routines (on large enough targets).