HardwareInfo: refactor and enrich
Besides device_name and device_version also determine
device_model and device_revision, and add retrieval functions
get_device_model() and ger_device_revision() for this data.
Another new fcuntion is get_device_string() that returns a
verbose string with most of the hw info.

Caching check did not work in the previous version but works
fine now.

Determining device model name uses a config file
(normally in /usr/share/enigma2/hw_info/hw_info.cfg) to define
how to scan the files in /proc/stb/info and if the data in that
file must be ammended.
So if a new box adds a new filename in /proc/stb/info to list the
receiver model name, this filename must be added to the config file.