Standby: Avoid GSOD when tunerMask is not initialized
This could happen when you watching a recording and the box goes to
standby and then after a timeout to deepstandby. This might still shut
down the box in the situation another box is streaming. I have to find a
better solution to check the tunerusage later but for now I at least avoid
this GSOD