Fix bug #220 - Movie Player won't create Deleted Items folder
More user-friendly handling error popup when
Tools.Trashcan.createTrashFolder() returns None.

Popups now have a more user-friendly message, and in the case of a
delete of a recording (rather than a directory), indicates that the
yes/no choice in this case will permanantly delete the recording.

If a directory is being deleted, and there's no .Trash, it now gives
an error message instead of silently ignoring the request. A fix
for this needs some more work.