Fix bug #220 - Movie Player won't create Deleted Items folder
If a directory is being moved to trash, and there's no .Trash, offer
the user the opportunity to delete it. Required a new method for
directory deletion for a reasonably clean implmentation.

Change the direct call of self.delete([True]) to self.delete(True) - there's
no need of a list in this context.