Fix Bug #243  -  Name-based search in Movie Player can cause crash
The list entry for '..' has an info element that is None, and the
original code skips over the '..' entry by starting the wraparound
search at the second item in the list. This only works if '..' is
the first list item. If Deleted Items is in the list, the wraparound
search starts with the second list entry and crashes when it accesses
the list entry for '..'.

Now search all items in the list, but skip the comparison with the
user string if the info entry in the list element is None.

Also fix a bug in getShortName() where the return value is only
uppercased if the serviceref was a directory. The caller compensated
for this by calling upper() on the return value of getShortName().
getShortName() now always returns upper case, and the upper() call
on the call of getShortName() has been removed.